2_synthetic phantom reconstruction/requirements.txt
absl-py==0.13.0
appnope==0.1.0
asn1crypto==0.24.0
astor==0.7.1
astropy==3.1.2
astunparse==1.6.3
atomicwrites==1.3.0
attrs==18.2.0
backcall==0.1.0
bleach==3.1.0
bokeh==1.0.4
Bottleneck==1.2.1
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.12.1
chardet==3.0.4
cleanlab==0.1.0
Click==7.0
cloudpickle==0.8.0
cryptography==2.5
cycler==0.10.0
cytoolz==0.9.0.1
dask==1.1.2
decorator==4.3.2
dill==0.2.8.2
distributed==1.26.0
entrypoints==0.3
fast-histogram==0.5
fastcache==1.1.0
flatbuffers==1.12
future==0.18.2
gast==0.2.2
glue-core==0.14.1
glue-vispy-viewers==0.11
gluoncv==0.5.0
gmpy2==2.0.8
google-auth==1.32.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.34.1
h5py==2.10.0
heapdict==1.0.0
hyperas==0.4.1
hyperopt==0.1.2
idna==2.6
imageio==2.5.0
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10
joblib==0.16.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.0
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.0.1
locket==0.2.0
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.1
more-itertools==5.0.0
mpl-scatter-density==0.4
mpmath==1.1.0
msgpack==0.6.1
mxnet==1.5.1.post0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.2
notebook==5.7.4
numpy==1.19.5
oauthlib==3.1.1
olefile==0.46
opt-einsum==3.3.0
packaging==19.0
pandas==0.24.1
pandocfilters==1.4.2
parso==0.3.4
partd==0.3.9
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
plotly==3.6.1
pluggy==0.9.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.17.3
psutil==5.5.0
ptyprocess==0.6.0
py==1.8.0
py4j==0.10.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pydot==1.4.1
pydot-ng==2.0.0
pydotplus==2.0.2
Pygments==2.3.1
pygraphviz==1.3
pymongo==3.7.2
PyOpenGL==3.1.1a1
pyOpenSSL==19.0.0
pyparsing==2.3.1
PySocks==1.6.8
pyspark==2.3.2
pytest==4.3.0
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.2.0
pytest-openfiles==0.3.2
pytest-remotedata==0.3.1
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
QtPy==1.6.0
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
scikit-image==0.15.0
scikit-learn==0.23.2
scipy==1.4.1
Send2Trash==1.5.0
six==1.15.0
sortedcontainers==2.1.0
sympy==1.4
TBB==0.1
tblib==1.3.2
tensorboard==2.1.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
Theano==1.0.5
threadpoolctl==2.1.0
tifffile==0.15.1
toolz==0.9.0
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
typing-extensions==3.7.4.3
urllib3==1.22
vtk==8.1.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.12.1
xlrd==1.2.0
zict==0.1.3
